ArchivesSpace DIP Upload tweaks
refs #9093

This commit sets the publish and restrictions flags on both digital objects
and file versions.

Also, notes are copied from the parent archival object to the digital object.
These new notes are converted to type 'general note' if the original is a
note type that can't be used with a digital object.

Sometimes a DIP does not include access derivatives,
for example when the user supplies access copies
(manual normalization).  Those access copies are
listed in the Files table as 'original'

This commit sets the digital object format and size and title
based on the original in these cases.