Revamp connection details support
Better support for `host` values that are URLs. Added some tests to make
very explicit the pairs that we are accepting and refusing.

I've updated the dependencies described in `local.txt`. Tests are still
passing but it also fixes a dependency problem that was preventing the
environment from building - see build #477140036:
https://travis-ci.org/artefactual-labs/agentarchives/builds/477140036.

This is connected to https://github.com/archivematica/Issues/issues/409.